merge

+++ SurcoucheCpp/System.Text.RegularExpressions/Match/Match.h
@@ -33,7 +33,7 @@ namespace System
                 String ToString() const override;
                 Boolean Equals(const object* _obj) override;
                 Boolean Equals(const object& _obj) override;
-                size_t GetHashCode() const override;
+                Integer GetHashCode() const override;
 #pragma endregion override
 #pragma region operator
             public:
