+++ SurcoucheCpp/System/Box/ConfirmationBox/ConfirmationBox.h
@@ -26,7 +26,7 @@ namespace System
         String ToString() const override;
         Boolean Equals(const object* _obj) override;
         Boolean Equals(const object& _obj) override;
-        size_t GetHashCode() const override;
+        Integer GetHashCode() const override;
 #pragma endregion override
 #pragma region operator
         ConfirmationBox& operator=(const ConfirmationBox& _other);
