+++ SurcoucheCpp/System/Event/EventArgs/EventArgs.h
@@ -14,7 +14,7 @@ namespace System
     public:
         Boolean Equals(const object* _obj) override;
         Boolean Equals(const object& _obj) override;
-        size_t GetHashCode() const override;
+        Integer GetHashCode() const override;
 #pragma endregion override
     };
 }
