+++ SurcoucheCpp/System/Version/Version.h
@@ -31,7 +31,7 @@ namespace System
         String ToString() const override;
         Boolean Equals(const object* _obj) override;
         Boolean Equals(const object& _obj) override;
-        size_t GetHashCode() const override;
+        Integer GetHashCode() const override;
         Boolean Equals(const Version& _object) override;
 #pragma endregion override
 #pragma region operator
